pyproject.toml
[tool.poetry]
name = "mcp-registry-harvester"
version = "0.1.0"
description = "Crawl and normalize MCP registries into a canonical dataset with query, API, and clustering."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "mcp_harvest", from = "src" },
    { include = "harvest", from = "src" },
    { include = "ml", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
httpx = {version = "0.27.2", extras=["http2"]}
tenacity = "8.5.0"
pydantic = "2.8.2"
beautifulsoup4 = "4.12.3"
selectolax = "0.3.21"
markdown-it-py = "3.0.0"
pandas = "2.2.2"
numpy = "2.0.2"
scikit-learn = "1.5.1"
fastapi = "0.112.1"
uvicorn = "0.30.6"
python-dotenv = "1.0.1"
rich = "13.7.1"
typer = "0.12.5"
matplotlib = "3.9.1"
httpx-sse = "0.4.0"
requests-cache = "1.2.1"
tldextract = "5.1.2"
yarl = "1.9.4"
rapidfuzz = "3.9.7"
pyarrow = "17.0.0"
sentence-transformers = "2.7.0"
jinja2 = "3.1.4"
pyyaml = "6.0.2"
joblib = "1.4.2"
openai = "1.43.0"

[tool.poetry.scripts]
mcp-harvest = "mcp_harvest.cli:app"


[tool.poetry.group.dev.dependencies]
black = "24.8.0"
ruff = "0.6.9"
isort = "5.13.2"
mypy = "1.11.2"
pytest = "8.3.2"
pytest-cov = "5.0.0"
pre-commit = "3.8.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "B"]
lint.ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
